# cmd/api/app/controllers.go
// Code generated by goagen v1.2.0-dirty, DO NOT EDIT.
//
// API "p7": Application Controllers
//
// Command:
// $ goagen
// --design=github.com/markusklems/p7/cmd/api/design
// --out=$(GOPATH)/src/github.com/markusklems/p7/cmd/api
// --version=v1.2.0-dirty

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Encoder.Register(goa.NewGobEncoder, "application/gob", "application/x-gob")
	service.Encoder.Register(goa.NewXMLEncoder, "application/xml")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")
	service.Decoder.Register(goa.NewGobDecoder, "application/gob", "application/x-gob")
	service.Decoder.Register(goa.NewXMLDecoder, "application/xml")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// CSSController is the controller interface for the CSS actions.
type CSSController interface {
	goa.Muxer
	goa.FileServer
}

// MountCSSController "mounts" a CSS resource controller on the given service.
func MountCSSController(service *goa.Service, ctrl CSSController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/css/*filepath", ctrl.MuxHandler("preflight", handleCSSOrigin(cors.HandlePreflight()), nil))

	h = ctrl.FileHandler("/css/*filepath", "public/css")
	h = handleCSSOrigin(h)
	service.Mux.Handle("GET", "/css/*filepath", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "CSS", "files", "public/css", "route", "GET /css/*filepath")

	h = ctrl.FileHandler("/css/", "public/css/index.html")
	h = handleCSSOrigin(h)
	service.Mux.Handle("GET", "/css/", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "CSS", "files", "public/css/index.html", "route", "GET /css/")
}

// handleCSSOrigin applies the CORS response headers corresponding to the origin.
func handleCSSOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, OPTIONS")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// HealthController is the controller interface for the Health actions.
type HealthController interface {
	goa.Muxer
	Check(*CheckHealthContext) error
}

// MountHealthController "mounts" a Health resource controller on the given service.
func MountHealthController(service *goa.Service, ctrl HealthController) {
	initService(service)
	var h goa.Handler

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCheckHealthContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Check(rctx)
	}
	service.Mux.Handle("GET", "/p7/health/check", ctrl.MuxHandler("check", h, nil))
	service.LogInfo("mount", "ctrl", "Health", "action", "Check", "route", "GET /p7/health/check")
}

// ImgController is the controller interface for the Img actions.
type ImgController interface {
	goa.Muxer
	goa.FileServer
}

// MountImgController "mounts" a Img resource controller on the given service.
func MountImgController(service *goa.Service, ctrl ImgController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/img/*filepath", ctrl.MuxHandler("preflight", handleImgOrigin(cors.HandlePreflight()), nil))

	h = ctrl.FileHandler("/img/*filepath", "public/img")
	h = handleImgOrigin(h)
	service.Mux.Handle("GET", "/img/*filepath", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Img", "files", "public/img", "route", "GET /img/*filepath")

	h = ctrl.FileHandler("/img/", "public/img/index.html")
	h = handleImgOrigin(h)
	service.Mux.Handle("GET", "/img/", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Img", "files", "public/img/index.html", "route", "GET /img/")
}

// handleImgOrigin applies the CORS response headers corresponding to the origin.
func handleImgOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, OPTIONS")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// JsController is the controller interface for the Js actions.
type JsController interface {
	goa.Muxer
	goa.FileServer
}

// MountJsController "mounts" a Js resource controller on the given service.
func MountJsController(service *goa.Service, ctrl JsController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/js/*filepath", ctrl.MuxHandler("preflight", handleJsOrigin(cors.HandlePreflight()), nil))

	h = ctrl.FileHandler("/js/*filepath", "public/js")
	h = handleJsOrigin(h)
	service.Mux.Handle("GET", "/js/*filepath", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Js", "files", "public/js", "route", "GET /js/*filepath")

	h = ctrl.FileHandler("/js/", "public/js/index.html")
	h = handleJsOrigin(h)
	service.Mux.Handle("GET", "/js/", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Js", "files", "public/js/index.html", "route", "GET /js/")
}

// handleJsOrigin applies the CORS response headers corresponding to the origin.
func handleJsOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, OPTIONS")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// LambdaController is the controller interface for the Lambda actions.
type LambdaController interface {
	goa.Muxer
	Code(*CodeLambdaContext) error
	Create(*CreateLambdaContext) error
	Delete(*DeleteLambdaContext) error
	List(*ListLambdaContext) error
	Run(*RunLambdaContext) error
	Show(*ShowLambdaContext) error
	Update(*UpdateLambdaContext) error
	Watch(*WatchLambdaContext) error
}

// MountLambdaController "mounts" a Lambda resource controller on the given service.
func MountLambdaController(service *goa.Service, ctrl LambdaController) {
	initService(service)
	var h goa.Handler

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCodeLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Code(rctx)
	}
	service.Mux.Handle("GET", "/p7/lambdas/:lambda_id/actions/code", ctrl.MuxHandler("code", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Code", "route", "GET /p7/lambdas/:lambda_id/actions/code")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCreateLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		// Build the payload
		if rawPayload := goa.ContextRequest(ctx).Payload; rawPayload != nil {
			rctx.Payload = rawPayload.(*LambdaPayload)
		} else {
			return goa.MissingPayloadError()
		}
		return ctrl.Create(rctx)
	}
	service.Mux.Handle("POST", "/p7/lambdas", ctrl.MuxHandler("create", h, unmarshalCreateLambdaPayload))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Create", "route", "POST /p7/lambdas")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewDeleteLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Delete(rctx)
	}
	service.Mux.Handle("DELETE", "/p7/lambdas/:lambda_id", ctrl.MuxHandler("delete", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Delete", "route", "DELETE /p7/lambdas/:lambda_id")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewListLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.List(rctx)
	}
	service.Mux.Handle("GET", "/p7/lambdas", ctrl.MuxHandler("list", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "List", "route", "GET /p7/lambdas")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewRunLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Run(rctx)
	}
	service.Mux.Handle("GET", "/p7/lambdas/:lambda_id/actions/run", ctrl.MuxHandler("run", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Run", "route", "GET /p7/lambdas/:lambda_id/actions/run")
	service.Mux.Handle("POST", "/p7/lambdas/:lambda_id/actions/run", ctrl.MuxHandler("run", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Run", "route", "POST /p7/lambdas/:lambda_id/actions/run")
	service.Mux.Handle("PUT", "/p7/lambdas/:lambda_id/actions/run", ctrl.MuxHandler("run", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Run", "route", "PUT /p7/lambdas/:lambda_id/actions/run")
	service.Mux.Handle("DELETE", "/p7/lambdas/:lambda_id/actions/run", ctrl.MuxHandler("run", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Run", "route", "DELETE /p7/lambdas/:lambda_id/actions/run")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewShowLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Show(rctx)
	}
	service.Mux.Handle("GET", "/p7/lambdas/:lambda_id", ctrl.MuxHandler("show", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Show", "route", "GET /p7/lambdas/:lambda_id")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewUpdateLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		// Build the payload
		if rawPayload := goa.ContextRequest(ctx).Payload; rawPayload != nil {
			rctx.Payload = rawPayload.(*LambdaPayload)
		} else {
			return goa.MissingPayloadError()
		}
		return ctrl.Update(rctx)
	}
	service.Mux.Handle("PATCH", "/p7/lambdas/:lambda_id", ctrl.MuxHandler("update", h, unmarshalUpdateLambdaPayload))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Update", "route", "PATCH /p7/lambdas/:lambda_id")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewWatchLambdaContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Watch(rctx)
	}
	service.Mux.Handle("GET", "/p7/lambdas/:lambda_id/watch", ctrl.MuxHandler("watch", h, nil))
	service.LogInfo("mount", "ctrl", "Lambda", "action", "Watch", "route", "GET /p7/lambdas/:lambda_id/watch")
}

// unmarshalCreateLambdaPayload unmarshals the request body into the context request data Payload field.
func unmarshalCreateLambdaPayload(ctx context.Context, service *goa.Service, req *http.Request) error {
	payload := &lambdaPayload{}
	if err := service.DecodeRequest(req, payload); err != nil {
		return err
	}
	if err := payload.Validate(); err != nil {
		// Initialize payload with private data structure so it can be logged
		goa.ContextRequest(ctx).Payload = payload
		return err
	}
	goa.ContextRequest(ctx).Payload = payload.Publicize()
	return nil
}

// unmarshalUpdateLambdaPayload unmarshals the request body into the context request data Payload field.
func unmarshalUpdateLambdaPayload(ctx context.Context, service *goa.Service, req *http.Request) error {
	payload := &lambdaPayload{}
	if err := service.DecodeRequest(req, payload); err != nil {
		return err
	}
	if err := payload.Validate(); err != nil {
		// Initialize payload with private data structure so it can be logged
		goa.ContextRequest(ctx).Payload = payload
		return err
	}
	goa.ContextRequest(ctx).Payload = payload.Publicize()
	return nil
}

// PublicController is the controller interface for the Public actions.
type PublicController interface {
	goa.Muxer
	goa.FileServer
}

// MountPublicController "mounts" a Public resource controller on the given service.
func MountPublicController(service *goa.Service, ctrl PublicController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/", ctrl.MuxHandler("preflight", handlePublicOrigin(cors.HandlePreflight()), nil))

	h = ctrl.FileHandler("/", "public/html/index.html")
	h = handlePublicOrigin(h)
	service.Mux.Handle("GET", "/", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Public", "files", "public/html/index.html", "route", "GET /")
}

// handlePublicOrigin applies the CORS response headers corresponding to the origin.
func handlePublicOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, OPTIONS")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// SwaggerController is the controller interface for the Swagger actions.
type SwaggerController interface {
	goa.Muxer
	goa.FileServer
}

// MountSwaggerController "mounts" a Swagger resource controller on the given service.
func MountSwaggerController(service *goa.Service, ctrl SwaggerController) {
	initService(service)
	var h goa.Handler

	h = ctrl.FileHandler("/swagger-ui/*filepath", "public/swagger-ui/")
	service.Mux.Handle("GET", "/swagger-ui/*filepath", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Swagger", "files", "public/swagger-ui/", "route", "GET /swagger-ui/*filepath")

	h = ctrl.FileHandler("/swagger.json", "public/swagger/swagger.json")
	service.Mux.Handle("GET", "/swagger.json", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Swagger", "files", "public/swagger/swagger.json", "route", "GET /swagger.json")

	h = ctrl.FileHandler("/swagger-ui/", "public/swagger-ui/index.html")
	service.Mux.Handle("GET", "/swagger-ui/", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Swagger", "files", "public/swagger-ui/index.html", "route", "GET /swagger-ui/")
}